[Decomp][Script] Cohesion split: extract script action execution to ScriptAction.cpp
